added global states
